Ex1::calculator::calculator
Struct Calculator

pub struct Calculator {
    commands: VecDeque<char>,
    operation_mode: i8,
    data: Vec<Operand>,
    registers: HashMap<char, Operand>,
}

A programmable calculator according to the specification in https://github.com/mwage/programmiersprachen/blob/master/aufgabe1.pdf.

Fields
commands: VecDeque<char>

A stream of characters regarded as commands to be executed in sequential order.

operation_mode: i8

An integer used in controlling the interpretation of commands

data: Vec<Operand>

A stack holding integers, floating-point numbers and strings when evaluating expressions in post-fix notation.

registers: HashMap<char, Operand>

A set of 52 read-only registers named by letters A to Z and a to z, each holding a single integer, floating-point number or string. These are constants (code that exists when switching on the calculator)

Implementations

impl Calculator

pub fn new() -> Self

Creates a new instance of the calculator, initializes registers: Register a: Welcome message + executes register b Register b: Loop that queries for user input Register c: Computes the ultimate answer to life, the universe, and everything Register d: Reverses an input string Register e: Reverses individual words of an input string and counts words, letters, digits, white-spaces and special characters

pub fn turn_on(&mut self)

Turns on the calculator and queries for user input

fn execute_commands(&mut self)

The first character in the command stream, we call it command character, gets executed. On execution this character is removed from the command stream and the next character becomes executable. The kind of character and the operation mode together determine the operation to be executed.

fn integer_construction(&mut self, next_command: char) -> Result<(), String>

Handle next command character when in integer construction mode

Arguments

next_command Next command character on the command stream

fn decimal_place_construction( &mut self, next_command: char, ) -> Result<(), String>

Handle next command character when in decimal construction mode

Arguments

next_command Next command character on the command stream

fn string_construction(&mut self, next_command: char) -> Result<(), String>

Handle next command character when in string construction mode/// Handle next command character when in integer construction mode

Arguments

next_command Next command character on the command stream

fn execution(&mut self, next_command: char) -> Result<(), String>

Handle next command character when in execution mode

Arguments

next_command Next command character on the command stream